Nuke `TableConstructor`
